requirements.txt
pygame-ce
tweener
requests
pypresence